requirements.txt
cryptography>=41.0.0
pyperclip>=1.8.2
pillow>=10.0.0 